# community-backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ServiceResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"anyOf": [],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ServiceResponse_any-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_any-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DocumentInput": {
				"properties": {
					"documentName": {
						"type": "string"
					},
					"documentUrl": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"publicId": {
						"type": "string",
						"nullable": true
					},
					"deleteToken": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"documentName",
					"documentUrl",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FeedbackCreateRequest": {
				"properties": {
					"projectId": {
						"type": "string",
						"nullable": true
					},
					"mainMessage": {
						"type": "string",
						"nullable": true
					},
					"feedbackType": {
						"type": "string",
						"enum": [
							"positive",
							"negative",
							"suggestion",
							"concern",
							null
						],
						"nullable": true
					},
					"feedbackMethod": {
						"type": "string",
						"enum": [
							"text",
							"voice",
							"video"
						]
					},
					"suggestions": {
						"type": "string",
						"nullable": true
					},
					"followUpNeeded": {
						"type": "boolean"
					},
					"documents": {
						"items": {
							"$ref": "#/components/schemas/DocumentInput"
						},
						"type": "array"
					},
					"responderName": {
						"type": "string",
						"nullable": true
					},
					"otherFeedbackOn": {
						"type": "string"
					}
				},
				"required": [
					"feedbackMethod"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FeedbackReplyCreateRequest": {
				"properties": {
					"subject": {
						"type": "string",
						"nullable": true
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FeedbackUpdateRequest": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"mainMessage": {
						"type": "string"
					},
					"feedbackType": {
						"type": "string",
						"enum": [
							"positive",
							"negative",
							"suggestion",
							"concern"
						]
					},
					"feedbackMethod": {
						"type": "string",
						"enum": [
							"text",
							"voice",
							"video"
						]
					},
					"suggestions": {
						"type": "string"
					},
					"followUpNeeded": {
						"type": "boolean"
					},
					"documents": {
						"items": {
							"$ref": "#/components/schemas/DocumentInput"
						},
						"type": "array"
					},
					"responderName": {
						"type": "string"
					},
					"otherFeedbackOn": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"submitted",
							"Acknowledged",
							"Resolved",
							"Rejected"
						]
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DocumentCreateRequest": {
				"properties": {
					"projectId": {
						"type": "string"
					},
					"documentName": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"type": {
						"type": "string",
						"nullable": true
					},
					"addedAt": {
						"type": "string",
						"format": "date-time"
					},
					"documentUrl": {
						"type": "string",
						"nullable": true
					},
					"userId": {
						"type": "string"
					},
					"publicId": {
						"type": "string",
						"nullable": true
					},
					"deleteToken": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"documentName",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "community-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Bikorimana Saveur"
		}
	},
	"paths": {
...
		"/api/feedback": {
			"get": {
				"operationId": "GetFeedback",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [
					{
						"jwt": [
							"feedback:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"submitted",
								"Acknowledged",
								"Resolved",
								"Rejected"
							]
						}
					},
					{
						"in": "query",
						"name": "feedbackType",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"positive",
								"negative",
								"suggestion",
								"concern"
							]
						}
					},
					{
						"in": "query",
						"name": "projectId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "owner",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"me",
								"other"
							]
						}
					},
					{
						"in": "query",
						"name": "org",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"mine",
								"others",
								"all"
							]
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateFeedback",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [
					{
						"jwt": [
							"feedback:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FeedbackCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/feedback/{feedbackId}": {
			"get": {
				"operationId": "GetFeedbackById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					},
					"404": {
						"description": "Feedback not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [
					{
						"jwt": [
							"feedback:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "feedbackId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateFeedback",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [
					{
						"jwt": [
							"feedback:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "feedbackId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FeedbackUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteFeedback",
				"responses": {
					"204": {
						"description": "No Content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [
					{
						"jwt": [
							"feedback:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "feedbackId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/feedback/{feedbackId}/replies": {
			"get": {
				"operationId": "GetFeedbackReplies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [
					{
						"jwt": [
							"feedback:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "feedbackId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "AddFeedbackReply",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [
					{
						"jwt": [
							"feedback:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "feedbackId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FeedbackReplyCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/feedback/stats/summary": {
			"get": {
				"operationId": "GetFeedbackStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [
					{
						"jwt": [
							"feedback:read"
						]
					}
				],
				"parameters": []
			}
		},
		"/api/documents": {
			"get": {
				"operationId": "GetDocuments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "projectId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateDocument",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Documents"
				],
				"security": [
					{
						"jwt": [
							"project:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DocumentCreateRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}